+++ RayTracerChallenge/Vector.cs
@@ -2,6 +2,6 @@
 
 public class Vector : Tuple
 {
-    public Vector(float x, float y, float z) : base(x,y,z,0) {}
+    public Vector(double x, double y, double z) : base(x,y,z,0) {}
 }
